refactor: 변환 예측 기능 분리 및 유틸 추가
