+++ next-converter/jest.setup.js
@@ -1,1 +1,1 @@
-require('@testing-library/jest-dom/extend-expect');
+import '@testing-library/jest-dom/extend-expect';
